reverting ia bridge changes since openlibrary.php polyfilled

+++ openlibrary/core/lending.py
@@ -345,11 +345,12 @@ def _get_ia_loans_of_user(userid):
     ia_loans = ia_lending_api.find_loans(userid=userid)
     return [Loan.from_ia_loan(d) for d in ia_loans]
 
-def create_loan(identifier, user_key, book_key=None):
+def create_loan(identifier, resource_type, user_key, book_key=None):
     """Creates a loan and returns it.
     """
     ia_loan = ia_lending_api.create_loan(
          identifier=identifier,
+         format=resource_type,
          userid=user_key,
          ol_key=book_key)
 
@@ -384,6 +385,7 @@ def sync_loan(identifier, loan=NOT_INITIALIZED):
     loan_data = loan and dict(
         uuid=loan['uuid'],
         loaned_at=loan['loaned_at'],
+        resource_type=loan['resource_type'],
         ocaid=loan['ocaid'],
         book=loan['book'])
 
@@ -459,9 +461,16 @@ class Loan(dict):
         _uuid = uuid.uuid4().hex
         loaned_at = time.time()
 
-        loan_link = BOOKREADER_STREAM_URL_PATTERN.format(config_bookreader_host, identifier)
-        t_expiry = datetime.datetime.utcnow() + datetime.timedelta(days=BOOKREADER_LOAN_DAYS)
-        expiry = t_expiry.isoformat()
+        if resource_type == "bookreader":
+            resource_id = "bookreader:" + identifier
+            loan_link = BOOKREADER_STREAM_URL_PATTERN.format(config_bookreader_host, identifier)
+            t_expiry = datetime.datetime.utcnow() + datetime.timedelta(days=BOOKREADER_LOAN_DAYS)
+            expiry = t_expiry.isoformat()
+        else:
+            raise Exception('No longer supporting ACS borrows directly from Open Library. Please go to Archive.org')
+
+        if not resource_id:
+            raise Exception('Could not find resource_id for %s - %s' % (identifier, resource_type))
 
         key = "loan-" + identifier
         return Loan({
@@ -477,6 +486,8 @@ class Loan(dict):
             'resource_type': 'bookreader',
             'resource_id': 'bookreader:%s' % identifier,
             'loaned_at': loaned_at,
+            'resource_type': resource_type,
+            'resource_id': resource_id,
             'loan_link': loan_link,
         })
 
@@ -499,6 +510,7 @@ class Loan(dict):
 
         # For historic reasons, OL considers expiry == None as un-fulfilled
         # loan.
+
         expiry = data.get('until')
 
         d = {
@@ -510,8 +522,11 @@ class Loan(dict):
             'book': book_key,
             'ocaid': data['identifier'],
             'expiry': expiry,
+            'fulfilled': data['fulfilled'],
             'uuid': 'loan-{0}'.format(data['id']),
             'loaned_at': time.mktime(created.timetuple()),
+            'resource_type': data['format'],
+            'resource_id': data['resource_id'],
             'loan_link': data['loan_link'],
             'stored_at': 'ia'
         }
@@ -547,8 +562,11 @@ class Loan(dict):
 
     def return_loan(self):
         logger.info("*** return_loan ***")
-        self.delete()
-        return True
+        if self['resource_type'] == 'bookreader':
+            self.delete()
+            return True
+        else:
+            return False
 
     def delete(self):
         loan = dict(self, returned_at=time.time())
@@ -637,10 +655,21 @@ def update_loan_status(identifier):
     if loan is None or loan.get('from_ia'):
         return
 
-    if loan.is_expired():
-        loan.delete()
-        return
+    if loan['resource_type'] == 'bookreader':
+        if loan.is_expired():
+            loan.delete()
+            return
+    else:
+        acs4_loan = ACS4Item(identifier).get_loan()
+        if not acs4_loan and not loan.is_yet_to_be_fulfilled():
+            logger.info("%s: loan returned or expired or timedout, deleting...", identifier)
+            loan.delete()
+            return
 
+        if loan['expiry'] != acs4_loan['until']:
+            loan['expiry'] = acs4_loan['until']
+            loan.save()
+            logger.info("%s: updated expiry to %s", identifier, loan['expiry'])
 
 class ACS4Item(object):
     """Represents an item on ACS4 server.
@@ -666,7 +695,22 @@ class ACS4Item(object):
         """Returns the information about loan in the ACS4 server.
         """
         d = self.get_data() or {}
-        return
+        if not d.get('resources'):
+            return
+        for r in d['resources']:
+            if r['loans']:
+                loan = dict(r['loans'][0])
+                loan['resource_id'] = r['resourceid']
+                loan['resource_type'] = self._format2resource_type(r['format'])
+                return loan
+
+    def _format2resource_type(self, format):
+        formats = {
+            "application/epub+zip": "epub",
+            "application/pdf": "pdf"
+        }
+        return formats[format]
+
 
 class IA_Lending_API:
     """Archive.org waiting list API.
@@ -682,8 +726,8 @@ class IA_Lending_API:
     def find_loans(self, **kw):
         return self._post(method="loan.query", **kw).get('result', [])
 
-    def create_loan(self, identifier, userid, ol_key):
-        response = self._post(method="loan.create", identifier=identifier, format="bookreader", userid=userid, ol_key=ol_key)
+    def create_loan(self, identifier, userid, format, ol_key):
+        response = self._post(method="loan.create", identifier=identifier, userid=userid, format=format, ol_key=ol_key)
         if response['status'] == 'ok':
             return response['result']['loan']
